consolidates steps
